refactor the code, move the models, input, output selection to audio_models.py file
